+++ common/math/Vector.h
@@ -128,6 +128,13 @@ class Vector {
     return *this;
   }
 
+  Vector<T, Size>& operator+=(const T& other) {
+    for (int i = 0; i < Size; i++) {
+      m_data[i] += other;
+    }
+    return *this;
+  }
+
   Vector<T, Size> elementwise_multiply(const Vector<T, Size>& other) const {
     Vector<T, Size> result;
     for (int i = 0; i < Size; i++) {
