+++ decompiler/config/jak2/hacks.jsonc
@@ -495,7 +495,7 @@
     "(method 181 gator)": [
       2, 3, 7, 10, 11, 14, 15, 16, 17, 18, 19, 20, 21, 22, 25, 26, 30
     ],
-    "(method 10 simple-sprite-system)": [0, 1, 3, 4, 5, 6, 9, 11],
+    "(method 10 simple-sprite-system)": [0],
     "target-pilot-post": [0, 2, 4, 13, 16, 22, 27, 41],
     "(anon-function 0 ruins-obs)": [
       0, 5, 7, 12, 13, 15, 17, 23, 24, 33, 42, 45, 46, 53, 58
