+++ decompiler/config/jak2/label_types.jsonc
@@ -66,7 +66,7 @@
   "sprite-distort": [["L27", "vu-function"]],
   "sprite-glow": [
     ["L29", "vu-function"],
-    ["L27", "vector"]
+    ["L27", "dma-packet"]
   ],
   "debug": [
     ["L250", "(inline-array vector)", 32],
