+++ decompiler/config/jak2/type_casts.jsonc
@@ -10525,5 +10525,21 @@
     [8, "a3", "pal-electric-fan"],
     [15, "a3", "pal-electric-fan"],
     [22, "t0", "pal-electric-fan"]
+  ],
+  "sprite-glow-init-engine": [
+    [[8, 22], "a0", "dma-packet"],
+    [[32, 50], "a1", "dma-packet"],
+    [[52, 74], "a1", "dma-packet"],
+    [[76, 84], "a0", "dma-packet"],
+    [[85, 92], "v1", "dma-packet"]
+  ],
+  "sprite-glow-add-sprite": [
+    [[0, 33], "v1", "sprite-glow-cnt-template"]
+  ],
+  "sprite-glow-add-simple-sprite": [
+    [[0, 33], "v1", "sprite-glow-ref-template"]
+  ],
+  "add-shader-to-dma": [
+    [[6, 8], "v1", "(pointer uint32)"]
   ]
 }
