+++ game/graphics/opengl_renderer/sprite/GlowRenderer.cpp
@@ -1,0 +1,5 @@
+#include "GlowRenderer.h"
+
+void GlowRenderer::add_sprite(const SpriteGlowOutput& data) {}
+
+void GlowRenderer::flush(SharedRenderState* render_state, ScopedProfilerNode& prof) {}
