+++ game/graphics/opengl_renderer/sprite/SpriteRenderer.cpp
