+++ goal_src/jak2/engine/gfx/hw/gs.gc
@@ -79,6 +79,7 @@
   (signal 96)
   (finish 97)
   (label 98)
+  (hack 127)
   )
 
 (defenum gs-reg64
