+++ goal_src/jak2/engine/gfx/sprite/simple-sprite-h.gc
@@ -8,17 +8,18 @@
 ;; DECOMP BEGINS
 
 (deftype sprite-glow-data (structure)
-  ((position   vector     :inline :offset-assert   0)
-   (size-x     float              :offset         12)
-   (size-probe float              :offset-assert  16)
-   (z-offset   float              :offset-assert  20)
-   (rot-angle  float              :offset-assert  24)
-   (size-y     float              :offset-assert  28)
-   (color      rgbaf      :inline :offset-assert  32)
-   (fade-a     float              :offset-assert  48)
-   (fade-b     float              :offset-assert  52)
-   (tex-id     texture-id         :offset-assert  56)
-   (dummy      uint32             :offset-assert  60)
+  ((position   vector :inline :offset-assert 0)
+   (size-x     float          :offset 12)
+   (size-probe float          :offset-assert 16)
+   (z-offset   float          :offset-assert 20)
+   (rot-angle  float          :offset-assert 24)
+   (size-y     float          :offset-assert 28)
+   (color      rgbaf  :inline :offset-assert 32)
+   (fade-a     float          :offset-assert 48)
+   (fade-b     float          :offset-assert 52)
+   (tex-id     texture-id     :offset-assert 56)
+   (dummy      uint32         :offset-assert 60)
+   (quads      vector 4 :inline :offset 0)
    )
   :method-count-assert 10
   :size-assert         #x40
@@ -39,16 +40,16 @@
   )
 
 (deftype simple-sprite-system (structure)
-  ((count     int16             :offset-assert   0)
-   (max-count int16             :offset-assert   2)
-   (data      sprite-glow-data  :offset-assert   4)
+  ((count     int16   :offset-assert 0)
+   (max-count int16   :offset-assert 2)
+   (data      (inline-array sprite-glow-data) :offset-assert 4)
    )
   :method-count-assert 12
   :size-assert         #x8
   :flag-assert         #xc00000008
   (:methods
-    (add! (_type_ dma-buffer) none 9)
-    (simple-sprite-system-method-10 (_type_ dma-buffer) none 10)
+    (add! (_type_ sprite-glow-data) none 9)
+    (draw-all-sprites! (_type_ dma-buffer) none 10)
     (clear! (_type_) none 11)
     )
   )
