+++ goal_src/jak2/engine/gfx/sprite/sprite.gc
@@ -929,10 +929,9 @@
       ;; sprite glow
       ;;;;;;;;;;;;;;;;
 
-      ;; TODO: figure out what this is doing
-      ; (sprite-glow-init-engine dma-buff)
-      ; (sprite-glow-draw dma-buff)
-      ; (simple-sprite-system-method-10 *simple-sprite-system* dma-buff)
+      (sprite-glow-init-engine dma-buff)
+      (sprite-glow-draw dma-buff)
+      (draw-all-sprites! *simple-sprite-system* dma-buff)
       (let* ((v1-26 dma-buff)
              (pkt6 (the-as dma-packet (-> v1-26 base)))
              )
