+++ goal_src/jak2/levels/city/common/vehicle-effects.gc
@@ -271,7 +271,7 @@
         )
       (when (nonzero? *simple-sprite-system*)
         ;; added nonzero check
-        (add! *simple-sprite-system* (the-as dma-buffer (-> s3-1 1)))
+        (add! *simple-sprite-system* (the sprite-glow-data (-> s3-1 1)))
         )
       (forward-up->quaternion (the-as quaternion (-> s3-1 0)) arg1 (the-as vector (-> obj rbody state matrix)))
       (quaternion-rotate-local-x! (the-as quaternion (-> s3-1 0)) (the-as quaternion (-> s3-1 0)) 32768.0)
@@ -386,7 +386,7 @@
               (set! (-> s4-0 color x) 255.0)
               (set! (-> s4-0 color y) (rand-vu-float-range 192.0 255.0))
               (set! (-> s4-0 color w) (* f30-1 (rand-vu-float-range 16.0 18.0)))
-              (add! *simple-sprite-system* (the-as dma-buffer s4-0))
+              (add! *simple-sprite-system* s4-0)
               (let ((f0-21 (-> obj camera-dist2))
                     (f1-6 245760.0)
                     )
@@ -399,7 +399,7 @@
                   (set! (-> s4-0 fade-b) 2.3332994)
                   (set! (-> s4-0 color z) (rand-vu-float-range 128.0 160.0))
                   (set! (-> s4-0 color w) (* f30-1 (rand-vu-float-range 32.0 36.0)))
-                  (add! *simple-sprite-system* (the-as dma-buffer s4-0))
+                  (add! *simple-sprite-system* s4-0)
                   )
                 )
               )
@@ -419,7 +419,7 @@
               (set! (-> s4-1 rot-angle) (* 182.04445 (rand-vu-float-range -4.0 4.0)))
               (set! (-> s4-1 color y) (* 64.0 (rand-vu)))
               (set! (-> s4-1 color w) (* f30-1 (rand-vu-float-range 16.0 21.0)))
-              (add! *simple-sprite-system* (the-as dma-buffer s4-1))
+              (add! *simple-sprite-system* s4-1)
               )
             )
           )
